@@ -1,11 +1,11 @@
 package academy.models;
 
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToMany;
-import javax.persistence.ManyToOne;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToMany;
+import jakarta.persistence.ManyToOne;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
 
@@ -15,7 +15,7 @@ public class Course {
     private String slug;
 
     private String name;
-    private String description;
+	private String description;
     private String examSlug;
 
     @ManyToOne(fetch = FetchType.LAZY)
