@@ -3,9 +3,9 @@ package academy.models;
 import java.time.LocalDate;
 import java.util.List;
 
-import javax.persistence.ElementCollection;
-import javax.persistence.Entity;
-import javax.persistence.Id;
+import jakarta.persistence.ElementCollection;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
 
 @Entity
 public class EntranceExam {
