@@ -2,7 +2,7 @@ package academy.models;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 public class SyllabusSection {
